Added sht40 drivers and reader; install script added and first time setup added to runAll.sh

+++ firmware/xu4Mqtt/pcbFirmware/orangepi/runAll.sh
@@ -1,6 +1,13 @@
 #!/bin/bash
 
+# /home/teamlary/canopy/firmware/xu4Mqtt/pcbFirmware/orangepi/.venv/bin/activate
+
 sleep 55
+
+# First time setup
+if [ ! -d "./.venv" ]; then
+    ./install.sh || exit 1
+fi
 
 kill $(pgrep -f 'python3 bme280Reader.py')
 sleep 5
@@ -13,13 +20,18 @@ sleep 5
 python3 cozIRReader.py &
 sleep 5
 
-
-kill $(pgrep -f 'python3 opcReader.py')  
-sleep 5
-python3 opcReader.py &
-sleep 5
+# OpcReader pending testing for Orange Pi Zero 2w
+#kill $(pgrep -f 'python3 opcReader.py')  
+#sleep 5
+#python3 opcReader.py &
+#sleep 5
 
 kill $(pgrep -f 'ips7100Reader.py')
 sleep 5
 python3 ips7100Reader.py &
+sleep 5
+
+kill $(pgrep -f 'sht40Reader.py')
+sleep 5
+python3 sht40Reader.py &
 sleep 5
